add RenderAirdrop component
