+++ app/components/airdrops/cards/DashBoardList.tsx
@@ -1,81 +1,11 @@
 "use client";
 import { Box, Button, Flex, Tab, TabList, Tabs, VStack } from "@chakra-ui/react";
 import { AddIcon } from "@chakra-ui/icons";
-import DashBoardCard from "./DashBoardCard";
 import { useTranslation } from "react-i18next";
-import { TemplateType } from "@/app/interfaces/dashboard";
-import { Airdrop } from "@/app/interfaces/dashboard";
-
-function RenderAirdrop() {
-  const airdrops: Airdrop[] = [
-    {
-      id: "asdf",
-      title: "Vesting",
-      templateName: TemplateType.LINEAR_VESTING,
-      owner: "0x",
-      tokenAddress: "0x",
-      createdAt: 1731897560,
-      merkleTreeUploadedAt: 1731897560,
-      contractAddress: "0x",
-      totalAirdropAmount: BigInt(50 * 1e18),
-      eligibleUsersNum: 1000,
-      claimedUsersNum: 40,
-      contractDeployedAt: 1731897560,
-      vestingEndsAt: 1731897560,
-    },
-    {
-      id: "zxcv",
-      title: "Standard",
-      templateName: TemplateType.STANDARD,
-      owner: "0x",
-      tokenAddress: "0x",
-      createdAt: 1731897560,
-      merkleTreeUploadedAt: undefined,
-      contractAddress: undefined,
-      totalAirdropAmount: BigInt(50 * 1e18),
-      eligibleUsersNum: 1000,
-      claimedUsersNum: 40,
-      contractDeployedAt: 1731897560,
-    },
-  ];
-  // return airdrops.map((airdrop) => {
-  //   <DashBoardCard key={airdrop.id} airdrop={airdrop} />;
-  // });
-}
+import RenderAirdrop from "./RenderAirdrop";
 
 export default function DashBoardList() {
   const { t } = useTranslation();
-  const airdrops: Airdrop[] = [
-    {
-      id: "asdf",
-      title: "Vesting",
-      templateName: TemplateType.LINEAR_VESTING,
-      owner: "0x",
-      tokenAddress: "0x",
-      createdAt: 1731897560,
-      merkleTreeUploadedAt: 1731897560,
-      contractAddress: "0x",
-      totalAirdropAmount: BigInt(0.005 * 1e18),
-      eligibleUsersNum: 1000,
-      claimedUsersNum: 40,
-      contractDeployedAt: 1731897560,
-      vestingEndsAt: 1731897560,
-    },
-    {
-      id: "zxcv",
-      title: "Standard",
-      templateName: TemplateType.STANDARD,
-      owner: "0x",
-      tokenAddress: "0x",
-      createdAt: 1731897560,
-      merkleTreeUploadedAt: undefined,
-      contractAddress: undefined,
-      totalAirdropAmount: BigInt(0.005 * 1e18),
-      eligibleUsersNum: 1000,
-      claimedUsersNum: 40,
-      contractDeployedAt: 1731897560,
-    },
-  ];
   return (
     <Box margin="0 auto" py={{ base: "3", md: "6" }}>
       <Flex justifyContent="space-between" alignItems="center" marginBottom="4">
@@ -101,9 +31,7 @@ export default function DashBoardList() {
         </Button>
       </Flex>
       <VStack spacing="4">
-        <DashBoardCard airdrop={airdrops[0]} />
-
-        <DashBoardCard airdrop={airdrops[1]} />
+        <RenderAirdrop />
       </VStack>
     </Box>
   );
